Rationalise command-line tools
- Move command-line tools to a dedicated command_line submodule.
- Use argparse subparsers for the various different image binning
  commands.
