+++ setup.cfg
@@ -47,7 +47,6 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    cues = tristan.cues:main
-    vds = tristan.vds:main
-    single-image = tristan.images:main_single_image
-    multi-image = tristan.images:main_multiple_images
+    cues = tristan.command_line.cues:main
+    vds = tristan.command_line.vds:main
+    images = tristan.command_line.images:main
