+++ src/tristan/command_line/cues.py
@@ -1,50 +1,25 @@
 """Summarise the cue messages in Tristan data."""
 
 import argparse
-import re
-import sys
 from contextlib import ExitStack
-from pathlib import Path
 
 import h5py
 import numpy as np
 from dask import array as da
 
-from . import cue_id_key, cue_keys, cue_time_key, cues, reserved, seconds
-from .data import data_files
+from .. import cue_id_key, cue_keys, cue_time_key, cues, reserved, seconds
+from ..data import data_files, find_input_file_name
+from . import input_parser, version_parser
 
-parser = argparse.ArgumentParser(description=__doc__)
-parser.add_argument(
-    "input_file",
-    help="Tristan raw data file ('.h5') file containing events data or detector "
-    "metadata.",
-    metavar="input-file",
+parser = argparse.ArgumentParser(
+    description=__doc__, parents=[version_parser, input_parser]
 )
-
-
-def find_file_name(in_file):
-    """Resolve the input file name."""
-    in_file = Path(in_file).expanduser().resolve()
-
-    data_dir = in_file.parent
-
-    # Get the segment 'name_root' from 'name_root_meta.h5' or 'name_root_000001.h5'.
-    file_name_root = re.fullmatch(r"(.*)_(?:meta|\d+)", in_file.stem)
-    if file_name_root:
-        file_name_root = file_name_root[1]
-    else:
-        sys.exit(
-            "Input file name did not have the expected format '<name>_meta.h5':\n"
-            f"\t{in_file}"
-        )
-
-    return data_dir, file_name_root
 
 
 def main(args=None):
     args = parser.parse_args(args)
 
-    data_dir, root = find_file_name(args.input_file)
+    data_dir, root = find_input_file_name(args.input_file)
     raw_files, _ = data_files(data_dir, root)
 
     with ExitStack() as stack:
